third_party/blink/web_tests/platform/generic/fast/js/function-name-is-in-scope-expected.txt

This tests verifies the scope of a function's name.

PASS: f.name == 'f' should be true and is.
PASS: f == 1 should be true and is.
PASS: g.name == 'g' should be true and is.
PASS: g == arguments.callee should be true and is.
